% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTables.R
\name{getTable}
\alias{getTable}
\title{Return a full database table}
\usage{
getTable(connection, tableName)
}
\arguments{
\item{connection}{connection variable returned from dbConnect()}

\item{tableName}{Name of table as shown in database}
}
\value{

}
\description{
Return a full database table
}
\examples{
getTable(conn, 'Decile')
}
