% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTables.R
\name{listDatabaseTables}
\alias{listDatabaseTables}
\title{Lists table names inside database}
\usage{
listDatabaseTables(connection, databaseName = "LearnerBDS", schema = "dbo")
}
\arguments{
\item{connection}{connection object returned from dbConnect()}

\item{databaseName}{name of database - defaults to 'LearnerBDS'}

\item{schema}{schema name - defaults to 'dbo'}
}
\value{

}
\description{
Lists table names inside database
}
\examples{
listDatabaseTables(conn)
listDatabaseTables(conn, databaseName="MyDatabase", schema="dbo")
}
