% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTables.R
\name{listMainObjects}
\alias{listMainObjects}
\title{List Main Catalog Objects}
\usage{
listMainObjects(connection)
}
\arguments{
\item{connection}{connection object returned from dbConnect()}
}
\value{

}
\description{
List Main Catalog Objects
}
\examples{
listMainObjects(conn)
}
