% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTables.R
\name{listTableColumnNames}
\alias{listTableColumnNames}
\title{Lists column names of table}
\usage{
listTableColumnNames(
  connection,
  tableName,
  databaseName = "LearnerBDS",
  schema = "dbo"
)
}
\arguments{
\item{connection}{connection object returned from dbConnect()}

\item{tableName}{name of table as shown in database}

\item{databaseName}{name of database - defaults to 'LearnerBDS'}

\item{schema}{schema name - defaults to 'dbo'}
}
\value{

}
\description{
Lists column names of table
}
\examples{
listTableColumnNames(conn, 'Gender')
listTableColumnNames(conn, 'MyTableName', databaseName='MyDatabase', schema='dbo')
}
