% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuery.R
\name{loadTableFromSqlFile}
\alias{loadTableFromSqlFile}
\title{Runs SQL query on database}
\usage{
loadTableFromSqlFile(connection, filepath)
}
\arguments{
\item{filepath}{location of the .sql file}

\item{conn}{connection object returned from dbConnect()}
}
\value{

}
\description{
Note: The filepath must contain forward slashes only ('/'), any back slashes ('\') should be converted to forward
}
\examples{
LoadSQLFile(conn, '/Queries/PerformanceScore_Subject.sql')
}
