% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuery.R
\name{runSQLQuery}
\alias{runSQLQuery}
\title{Runs SQL query on connection}
\usage{
runSQLQuery(connection, query)
}
\arguments{
\item{connection}{connection object returned from dbConnect()}

\item{query}{SQL query string}
}
\value{

}
\description{
Runs SQL query on connection
}
\examples{
runSQLQuery(
  conn,
  query = '
  SELECT * FROM SchoolLeaver
  WHERE SchoolLeaverYear >= 2016
  '
)

runSQLQuery(conn, queryString)

}
